ai-rca-poc/requirements.txt
pandas>=2.0.0
streamlit>=1.35.0
plotly>=5.20.0
openai>=1.30.0
python-dotenv>=1.0.0
pyyaml>=6.0
